refactor: move shared VL image helpers into mm_image
Kimi and Qwen3.5 were importing file:// loaders, part detectors, and
layout_model_name from qwen3_vl. Lift those cross-family helpers into
renderers.mm_image so family modules no longer depend on each other for
I/O utilities.
